+++ sixtyfps_runtime/interpreter/dynamic_component.rs
@@ -1355,9 +1355,7 @@ pub fn instantiate(
                     InstanceRef::from_pin_ref(c, guard)
                 }),
             );
-            sixtyfps_corelib::model::ModelHandle(Some(Rc::new(
-                crate::value_model::ValueModel::new(m),
-            )))
+            sixtyfps_corelib::model::ModelRc::new(Rc::new(crate::value_model::ValueModel::new(m)))
         });
     }