+++ sixtyfps_runtime/interpreter/ffi.rs
@@ -85,7 +85,7 @@ pub unsafe extern "C" fn sixtyfps_interpreter_value_new_array_model(
     let vec = std::mem::transmute::<SharedVector<ValueOpaque>, SharedVector<Value>>(a.clone());
     std::ptr::write(
         val as *mut Value,
-        Value::Model(ModelHandle::new(Rc::new(SharedVectorModel::<Value>::from(vec)))),
+        Value::Model(ModelRc::new(Rc::new(SharedVectorModel::<Value>::from(vec)))),
     )
 }
 
@@ -121,7 +121,7 @@ pub unsafe extern "C" fn sixtyfps_interpreter_value_new_model(
 ) {
     std::ptr::write(
         val as *mut Value,
-        Value::Model(ModelHandle::new(Rc::new(ModelAdaptorWrapper(model)))),
+        Value::Model(ModelRc::new(Rc::new(ModelAdaptorWrapper(model)))),
     )
 }